Add variables, fix battleclick2, add wait function, fix treasureclick2
Add new variables to look at run once for paintings or lab (not sure
which to choose yet) and if using party 2 is desired.

Added wait function to look for paintings after returning from a
previous painting.

Fixed battleclick2 and treasureclick2 that would run infinitely while
looking for certain images